+++ client/src/pages/ItemDetailsPage/styles.jsx
@@ -1,0 +1,30 @@
+import { makeStyles } from '@material-ui/core/styles';
+import { red } from '@material-ui/core/colors';
+
+
+const useStyles = makeStyles((theme) => ({
+  root: {
+    maxWidth: 345,
+    alignItems: 'center',
+    justifyContent: 'center'
+  },
+  media: {
+    height: 0,
+    paddingTop: '56.25%', // 16:9
+  },
+  expand: {
+    transform: 'rotate(0deg)',
+    marginLeft: 'auto',
+    transition: theme.transitions.create('transform', {
+      duration: theme.transitions.duration.shortest,
+    }),
+  },
+  expandOpen: {
+    transform: 'rotate(180deg)',
+  },
+  avatar: {
+    backgroundColor: red[500],
+  },
+}));
+
+export default useStyles;
